requirements.txt
kokoro
moviepy
numpy
Pillow
python-dotenv
google-generativeai
beautifulsoup4
InquirerPy
EbookLib
pydub
clint
requests
soundfile
